+++ core/src/Graphics/VertexArrayFace2d.cpp
@@ -25,8 +25,8 @@ static std::vector<Vector3d<float>> generateFaceNormals(const std::vector<Vector
 }
 
 static std::vector<Vector2d<float>> generateFaceUVCoords(const std::vector<Vector2d<float>> &graphicalShape,
-                                                         const Vector2d<float> &size,
-                                                         const Vector2d<float> &position,
+                                                         const Vector2d<int> &size,
+                                                         const Vector2d<int> &position,
                                                          const textureMode &texModeX,
                                                          const textureMode &texModeY,
                                                          const float z)
@@ -81,8 +81,8 @@ static std::vector<Vector2d<float>> generateFaceUVCoords(const std::vector<Vecto
 VertexArrayFace2d::VertexArrayFace2d(const std::vector<Vector2d<float>> &graphicalShape,
                                      const float z,
                                      const std::string &texturePath,
-                                     const Vector2d<float> &textureSize,
-                                     const Vector2d<float> &texturePos,
+                                     const Vector2d<int> &textureSize,
+                                     const Vector2d<int> &texturePos,
                                      const Vector2d<textureMode> &texMode)
     : VertexArray(generateFaceCoords(graphicalShape, z),
                   generateFaceUVCoords(graphicalShape, textureSize, texturePos, texMode.x, texMode.y, z),
